서버구축 - 3주차 피드백 - refactoring - 패키징 위치 조정
