+++ hhplus-server-java/src/main/java/kr/hhplus/be/server/application/account/AccountCommand.java
@@ -1,5 +1,6 @@
 package kr.hhplus.be.server.application.account;
 
+import kr.hhplus.be.server.domain.account.AccountInfo;
 import lombok.Builder;
 
 @Builder
