--- hhplus-server-java/src/main/java/kr/hhplus/be/server/application/account/AccountQueryService.java
@@ -1,4 +1,0 @@
-package kr.hhplus.be.server.application.account;
-
-public class AccountQueryService {
-}
