+++ hhplus-server-java/src/main/java/kr/hhplus/be/server/application/order/OrderCommand.java
@@ -1,5 +1,6 @@
 package kr.hhplus.be.server.application.order;
 
+import kr.hhplus.be.server.domain.order.OrderAddressInfo;
 import lombok.Builder;
 
 import java.util.Map;
