+++ hhplus-server-java/src/main/java/kr/hhplus/be/server/domain/coupon/CouponApplyInfo.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.application.coupon;
+package kr.hhplus.be.server.domain.coupon;
 
 import lombok.Builder;
