+++ hhplus-server-java/src/main/java/kr/hhplus/be/server/domain/coupon/CouponIssueInfo.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.application.coupon;
+package kr.hhplus.be.server.domain.coupon;
 
 import kr.hhplus.be.server.domain.coupon.entity.CouponIssue;
 import lombok.Builder;
