+++ hhplus-server-java/src/main/java/kr/hhplus/be/server/domain/order/OrderAddressInfo.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.application.order;
+package kr.hhplus.be.server.domain.order;
 
 import java.time.LocalDateTime;
