+++ hhplus-server-java/src/main/java/kr/hhplus/be/server/domain/payment/PaymentRepository.java
@@ -1,6 +1,5 @@
 package kr.hhplus.be.server.domain.payment;
 
-import kr.hhplus.be.server.application.payment.PaymentInfo;
 import kr.hhplus.be.server.domain.payment.entity.Payment;
 
 import java.util.Optional;
