+++ hhplus-server-java/src/main/java/kr/hhplus/be/server/infra/payment/PaymentRepositoryImpl.java
@@ -1,6 +1,6 @@
 package kr.hhplus.be.server.infra.payment;
 
-import kr.hhplus.be.server.application.payment.PaymentInfo;
+import kr.hhplus.be.server.domain.payment.PaymentInfo;
 import kr.hhplus.be.server.domain.payment.PaymentRepository;
 import kr.hhplus.be.server.domain.payment.entity.Payment;
 import org.springframework.stereotype.Repository;
