+++ hhplus-server-java/src/main/java/kr/hhplus/be/server/interfaces/coupon/CouponIssueRequestDTO.java
@@ -1,7 +1,7 @@
 package kr.hhplus.be.server.interfaces.coupon;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import kr.hhplus.be.server.application.coupon.CouponIssueCommand;
+import kr.hhplus.be.server.domain.coupon.CouponIssueCommand;
 import lombok.Builder;
